requirements.txt
python-dotenv>=1.0.1
rich>=13.9.0
httpx>=0.28.0
